refactor(machine): unify maxCallDepth/callDepth as int (Finding 4 Stage 1)

+++ engine.go
@@ -41,8 +41,9 @@ var ErrEngineClosed = werr.NewStaticError("engine is closed")
 
 // DefaultMaxCallDepth is the default call depth limit for new engines.
 // At ~500 bytes per frame, 10000 frames ≈ 5MB. Use WithMaxCallDepth(0)
-// to opt out of the limit explicitly.
-const DefaultMaxCallDepth uint64 = 10000
+// to opt out of the limit explicitly; WithMaxCallDepth(n) with n < 0 is
+// clamped to 0 (also unlimited).
+const DefaultMaxCallDepth int = 10000
 
 // Engine is the main entry point for embedding Wile.
 //
@@ -61,7 +62,7 @@ type Engine struct {
 	lastCounters        machine.VMCounters
 	closers             []registry.Closeable
 	closed              bool
-	maxCallDepth        uint64
+	maxCallDepth        int
 	maxStackSize        uint64
 	inlineThreshold     int
 	contractEnforcement bool // propagated to RegisterPrimitive via cfg
